refactor: Ranking 추출 calculator클래스 제거
