+++ src/main/java/lotto/Application.java
@@ -2,13 +2,11 @@ package lotto;
 
 import lotto.controller.LottoController;
 import lotto.service.LottoAutoGenerator;
-import lotto.service.LottoRankingCalculator;
 
 public class Application {
 
     public static void main(String[] args) {
-        LottoController lottoController = new LottoController(new LottoAutoGenerator(),
-            LottoRankingCalculator.getInstance());
+        LottoController lottoController = new LottoController(new LottoAutoGenerator());
         lottoController.run();
     }
 }
